Update Android mode; make build reproducible; switch to AAPT2.
Honestly way too many things in one commit. Unfortunately it was easiest to do these things all at the same time.
Supporting Android 12 required a JNI version of AAPT; this was easiest to achieve by moving to AAPT2. Moving to
AAPT2 required updating the sketch previewer (because AAPT-generated resources appear to be incompatible with
AAPT2-generated resources). It made sense to make the artifacts build reproducible while updating the sketch
previewer. Upgrading Android mode required moving to androidx; it made sense to do this at the same time as
moving all of the asset generation to gradle. You may question my life choices, but I believe it is done now.

All of the derived assets (i.e., assets taken from other things, like desktop Android mode or the support
libraries) are now built in a fully reproducible way, instead of the previous ad-hoc and hard to maintain system.
We have a bunch of gradle tasks for downloading the releases of Android mode, Processing, the support libraries,
and even the Android platform. This way updating Android mode is, ideally, as simple as changing the upstream
URL to pull from. Note that this is my first foray into writing real Gradle tasks so they might be kind of gross.
The gradle scripts live in gradle/.

The androidx dependencies are pulled in from maven. I refer to them as artifacts. Additionally, the Android core,
VR, and AR libraries are pulled in through this artifact system. This new approach is much more modular and
allows us to more easily add targets and dependencies in the future. Another note is that we are now including
the transitive set of all androidx dependencies, whereas in the past I cherry-picked only what was needed for
builds to succeed. This is obviously the correct thing to do, but it probably makes the APK larger.

Adding support for AR sketches is probably a minimal amount of effort at this point.